requirements.txt
flask==1.1.2
request==2.25.1